Move auth.service into the shared services folder.
