+++ stem-explorer-ng/src/app/containers/register-page/register-page.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { AuthService } from 'src/app/auth.service';
+import { AuthService } from 'src/app/shared/services/auth.service';
 
 
 @Component({
