+++ stem-explorer-ng/src/app/shared/services/auth.service.ts
